# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tutor-support-system",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    },
    {
      "url": "https://open.oapi.vn"
    },
    {
      "url": "https://api.cloudinary.com"
    }
  ],
  "paths": {
...
    "/friends/": {
      "get": {
        "summary": "GetAllFriends",
        "tags": [],
        "responses": {},
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/conversations/": {
      "post": {
        "summary": "CreateConversation",
        "tags": [],
        "responses": {},
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "memberIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "GetConversation",
        "tags": [],
        "responses": {},
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}